Refactor: Moved Game Object Ordering functions into a new separate file
