+++ src/components/pages/EditionsPage.tsx
@@ -16,7 +16,7 @@ import { exportSingleEdition } from "../codeGeneration/Card/index";
 import type { Rule } from "../ruleBuilder/types";
 import { EditionData, slugify } from "../data/BalatroUtils";
 import { UserConfigContext } from "../Contexts";
-import { updateGameObjectIds, getObjectName } from "./JokersPage";
+import { updateGameObjectIds, getObjectName } from "../generic/GameObjectOrdering";
 
 
 interface EditionsPageProps {
