+++ src/components/pages/enhancements/EditEnhancementInfo.tsx
@@ -18,7 +18,7 @@ import {
 } from "../../generic/validationUtils";
 import { applyAutoFormatting } from "../../generic/balatroTextFormatter";
 import { UserConfigContext } from "../../Contexts";
-import { getObjectName, updateGameObjectIds } from "../JokersPage";
+import { updateGameObjectIds, getObjectName } from "../../generic/GameObjectOrdering";
 
 
 interface EditEnhancementInfoProps {
