+++ src/components/pages/enhancements/EnhancementCard.tsx
@@ -21,7 +21,7 @@ import Tooltip from "../../generic/Tooltip";
 import { formatBalatroText } from "../../generic/balatroTextFormatter";
 import { validateJokerName } from "../../generic/validationUtils";
 import { EnhancementData, slugify } from "../../data/BalatroUtils";
-import { getObjectName,updateGameObjectIds } from "../JokersPage";
+import { updateGameObjectIds, getObjectName } from "../../generic/GameObjectOrdering";
 
 interface EnhancementCardProps {
   enhancement: EnhancementData;
