+++ src/components/pages/seals/SealCard.tsx
@@ -16,7 +16,7 @@ import Tooltip from "../../generic/Tooltip";
 import { formatBalatroText } from "../../generic/balatroTextFormatter";
 import { validateJokerName } from "../../generic/validationUtils";
 import { SealData, slugify } from "../../data/BalatroUtils";
-import { getObjectName, updateGameObjectIds } from "../JokersPage";
+import { updateGameObjectIds, getObjectName } from "../../generic/GameObjectOrdering";
 
 interface SealCardProps {
   seal: SealData;
